requirements.txt
flask==2.2.3
neo4j==5.5.0
transformers==4.28.0
torch==1.13.1
sentence-transformers==2.2.2
numpy==1.24.1
scipy==1.10.0
python-dotenv==1.0.0
pytesseract==0.3.10
PyPDF2==3.0.0
SpeechRecognition==3.9.0
werkzeug==2.2.3
openai==0.27.0
huggingface-hub==0.12.1
nltk==3.8.1
regex==2023.6.3
matplotlib==3.6.3
networkx==3.1
requests==2.31.0
gunicorn==20.1.0
tensorflow==2.11.0
keras==2.11.0
tf-keras